♻️ [Refactor] OAuth 콜백URL 변경
기존 콜백URL은 /New였지만 /로 변경함에 따라 로직 위치 이동
